requirements.txt
apipkg==1.4
boto3==1.4.6
botocore==1.6.3
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.33.0
django-bootstrap4==0.0.1
django-dotenv==1.4.1
django-filter==1.0.4
django-invitations==1.9.1
django-pure-pagination==0.3.0
django-s3direct==1.0.1.dev1
django-storages==1.6.5
djangorestframework==3.6.4
docutils==0.14
execnet==1.4.1
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
Markdown==2.6.9
oauthlib==2.0.2
psycopg2==2.7.3
py==1.4.34
pytest==3.2.2
pytest-forked==0.2
pytest-xdist==1.20.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.0
coverage==4.5.1
